# supervisor-console/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zippy Logitech API",
    "description": "Logistics Business Intelligence Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pricing/ml-estimate": {
      "post": {
        "tags": [
          "ML Pricing"
        ],
        "summary": "Ml Price Estimate",
        "description": "ML-enhanced dynamic pricing estimate.\n\nUses LightGBM surge prediction (when model is trained) or rule-based fallback.\nProduces detailed cost breakdown with Indian logistics specifics:\n- City tier pricing (Metro discount, Rural surcharge)\n- Festival/remote/hill/congestion surcharges\n- Demand-supply dynamic surge\n- Service type multiplier (standard/express/priority)\n- Customer tier adjustments\n- Platform fee (3-5%) vs broker (8-12%)\n- Route Difficulty Score surcharge\n- Urbanization density multiplier\n- Deadhead / lane viability multiplier\n- GST estimate, with final tax treatment owned by the finance/GST layer",
        "operationId": "ml_price_estimate_api_v1_pricing_ml_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLPricingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pricing/surge-predict": {
      "post": {
        "tags": [
          "ML Pricing"
        ],
        "summary": "Predict Surge",
        "description": "Predict demand-supply surge multiplier.\n\nReturns surge multiplier (1.0-3.0) with confidence score.\nUses LightGBM model when trained, rule-based fallback otherwise.",
        "operationId": "predict_surge_api_v1_pricing_surge_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurgePredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pricing/rate-card": {
      "get": {
        "tags": [
          "ML Pricing"
        ],
        "summary": "Get Detailed Rate Card",
        "description": "Get current dynamic rate card with all multipliers and surcharges.\n\nIncludes base rates, city tier multipliers, surcharge ranges,\nservice multipliers, platform fee structure, and GST rates.",
        "operationId": "get_detailed_rate_card_api_v1_pricing_rate_card_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pricing/compare": {
      "post": {
        "tags": [
          "ML Pricing"
        ],
        "summary": "Compare Pricing",
        "description": "Compare Zippy pricing vs traditional broker pricing.\n\nShows side-by-side cost breakdown demonstrating 18-25% savings.",
        "operationId": "compare_pricing_api_v1_pricing_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/shipments/status": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "summary": "Get Shipments Status",
        "description": "Return role-safe shipment status projections for frontend tracking views.",
        "operationId": "get_shipments_status_api_v1_shipments_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter status by order id",
              "title": "Order Id"
            },
            "description": "Filter status by order id"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentStatusListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MLPricingRequest": {
        "properties": {
          "weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg",
            "description": "Cargo weight in kg"
          },
          "distance_km": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Distance Km",
            "description": "Distance in km"
          },
          "vehicle_category": {
            "type": "string",
            "title": "Vehicle Category",
            "description": "LCV, ICV, HCV, Tipper, Tractor, Trailer",
            "default": "HCV"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City",
            "description": "Origin city for tier detection",
            "default": ""
          },
          "destination_city": {
            "type": "string",
            "title": "Destination City",
            "description": "Destination city for RDS",
            "default": ""
          },
          "is_interstate": {
            "type": "boolean",
            "title": "Is Interstate",
            "default": false
          },
          "is_festival": {
            "type": "boolean",
            "title": "Is Festival",
            "default": false
          },
          "is_remote": {
            "type": "boolean",
            "title": "Is Remote",
            "default": false
          },
          "is_hill": {
            "type": "boolean",
            "title": "Is Hill",
            "default": false
          },
          "is_congested": {
            "type": "boolean",
            "title": "Is Congested",
            "default": false
          },
          "service_type": {
            "type": "string",
            "title": "Service Type",
            "description": "standard, express, priority",
            "default": "standard"
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type",
            "description": "high_value, medium, low_value",
            "default": "medium"
          },
          "demand": {
            "type": "integer",
            "title": "Demand",
            "description": "Current demand (number of orders)",
            "default": 10
          },
          "supply": {
            "type": "integer",
            "title": "Supply",
            "description": "Available vehicles in area",
            "default": 5
          },
          "diesel_price": {
            "type": "number",
            "title": "Diesel Price",
            "description": "Current diesel price INR/litre",
            "default": 89.62
          },
          "route_difficulty_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Difficulty Score",
            "description": "Optional 0-100 route difficulty score override"
          },
          "lane_viability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lane Viability",
            "description": "highly_balanced, moderately_balanced, unbalanced_origin_heavy, seasonal, remote_low_demand"
          },
          "return_load_probability": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Load Probability",
            "description": "Optional probability of finding a return load on the lane"
          }
        },
        "type": "object",
        "required": [
          "weight_kg",
          "distance_km"
        ],
        "title": "MLPricingRequest"
      },
      "MLPricingResponse": {
        "properties": {
          "base_cost": {
            "type": "number",
            "title": "Base Cost"
          },
          "base_rate_per_km": {
            "type": "number",
            "title": "Base Rate Per Km"
          },
          "city_tier": {
            "type": "integer",
            "title": "City Tier"
          },
          "tier_multiplier": {
            "type": "number",
            "title": "Tier Multiplier"
          },
          "fuel_index": {
            "type": "number",
            "title": "Fuel Index"
          },
          "tier_adjusted_cost": {
            "type": "number",
            "title": "Tier Adjusted Cost"
          },
          "route_difficulty": {
            "type": "object",
            "title": "Route Difficulty"
          },
          "urbanization_density": {
            "type": "object",
            "title": "Urbanization Density"
          },
          "surcharges": {
            "type": "object",
            "title": "Surcharges"
          },
          "total_surcharge_pct": {
            "type": "number",
            "title": "Total Surcharge Pct"
          },
          "surcharge_amount": {
            "type": "number",
            "title": "Surcharge Amount"
          },
          "density_adjusted_cost": {
            "type": "number",
            "title": "Density Adjusted Cost"
          },
          "surge_multiplier": {
            "type": "number",
            "title": "Surge Multiplier"
          },
          "surge_model": {
            "type": "string",
            "title": "Surge Model"
          },
          "surge_confidence": {
            "type": "number",
            "title": "Surge Confidence"
          },
          "surge_amount": {
            "type": "number",
            "title": "Surge Amount"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "service_multiplier": {
            "type": "number",
            "title": "Service Multiplier"
          },
          "service_amount": {
            "type": "number",
            "title": "Service Amount"
          },
          "deadhead_lane_viability": {
            "type": "object",
            "title": "Deadhead Lane Viability"
          },
          "deadhead_amount": {
            "type": "number",
            "title": "Deadhead Amount"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type"
          },
          "customer_adjustment": {
            "type": "number",
            "title": "Customer Adjustment"
          },
          "platform_fee_pct": {
            "type": "number",
            "title": "Platform Fee Pct"
          },
          "platform_fee": {
            "type": "number",
            "title": "Platform Fee"
          },
          "gst_breakdown": {
            "type": "object",
            "title": "Gst Breakdown"
          },
          "gst_amount": {
            "type": "number",
            "title": "Gst Amount"
          },
          "savings_vs_broker_pct": {
            "type": "number",
            "title": "Savings Vs Broker Pct"
          },
          "final_price": {
            "type": "number",
            "title": "Final Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "INR"
          },
          "breakdown_per_km": {
            "type": "number",
            "title": "Breakdown Per Km"
          }
        },
        "type": "object",
        "required": [
          "base_cost",
          "base_rate_per_km",
          "city_tier",
          "tier_multiplier",
          "fuel_index",
          "tier_adjusted_cost",
          "route_difficulty",
          "urbanization_density",
          "surcharges",
          "total_surcharge_pct",
          "surcharge_amount",
          "density_adjusted_cost",
          "surge_multiplier",
          "surge_model",
          "surge_confidence",
          "surge_amount",
          "service_type",
          "service_multiplier",
          "service_amount",
          "deadhead_lane_viability",
          "deadhead_amount",
          "subtotal",
          "customer_type",
          "customer_adjustment",
          "platform_fee_pct",
          "platform_fee",
          "gst_breakdown",
          "gst_amount",
          "savings_vs_broker_pct",
          "final_price",
          "breakdown_per_km"
        ],
        "title": "MLPricingResponse"
      },
...
      "ShipmentStatusListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "shipments": {
            "items": {
              "$ref": "#/components/schemas/ShipmentStatusResponse"
            },
            "type": "array",
            "title": "Shipments"
          }
        },
        "type": "object",
        "required": [
          "total",
          "shipments"
        ],
        "title": "ShipmentStatusListResponse"
      },
      "ShipmentStatusResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "shipment_status": {
            "type": "string",
            "title": "Shipment Status"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City"
          },
          "destination_city": {
            "type": "string",
            "title": "Destination City"
          },
          "latest_milestone": {
            "type": "string",
            "title": "Latest Milestone"
          },
          "delay_risk": {
            "type": "string",
            "title": "Delay Risk"
          },
          "current_eta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Eta"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "shipment_status",
          "origin_city",
          "destination_city",
          "latest_milestone",
          "delay_risk",
          "current_eta",
          "customer_phone",
          "updated_at"
        ],
        "title": "ShipmentStatusResponse"
      },
...
      "SurgePredictRequest": {
        "properties": {
          "demand": {
            "type": "integer",
            "title": "Demand",
            "default": 10
          },
          "supply": {
            "type": "integer",
            "title": "Supply",
            "default": 5
          },
          "city": {
            "type": "string",
            "title": "City",
            "default": ""
          },
          "is_remote": {
            "type": "boolean",
            "title": "Is Remote",
            "default": false
          },
          "is_hill": {
            "type": "boolean",
            "title": "Is Hill",
            "default": false
          },
          "is_festival": {
            "type": "boolean",
            "title": "Is Festival",
            "default": false
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km",
            "default": 500
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City",
            "default": ""
          },
          "destination_city": {
            "type": "string",
            "title": "Destination City",
            "default": ""
          },
          "congestion_level": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Congestion Level",
            "default": 0.3
          },
          "vehicle_category": {
            "type": "string",
            "title": "Vehicle Category",
            "default": "HCV"
          },
          "vehicle_age": {
            "type": "integer",
            "title": "Vehicle Age",
            "default": 3
          },
          "diesel_price": {
            "type": "number",
            "title": "Diesel Price",
            "default": 89.62
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type",
            "default": "medium"
          }
        },
        "type": "object",
        "title": "SurgePredictRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}